ebook/requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
idna==2.8
img==2.5
jieba==0.39
oauthlib==3.0.1
Pillow==5.4.1
pycryptodomex==3.7.2
python-alipay-sdk==1.10.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
